Add gazelle rule that builds a script to run Gazelle (#2)
This is very similar to the gazelle rule in rules_go with some small
improvements. This script may be copied into the workspace and run
without Bazel. When run in this fashion, the script will rebuild
itself (to check whether it's out of date), then it will build and run
Gazelle.

Also, several infrastructure changes:

* Added deps.bzl so that other repos can load Gazelle and the gazelle rule.
* Added an alias at //cmd/gazelle:gazelle. The executable will be
  moved here shortly.
* Updated setup documentation in README.rst.