cmd/gazelle: add -experimental_out_dir flag
This flag allows users to specify a directory where build files will
be written. There is also -experimental_out_suffix, which adds a
suffix to generated build file names.

This will be used for experimenting with new repository rules. The
flags will be removed at some point, but the functionality will
remain.