Rewrite packages.PlatformStrings (again) (#38)
PackageStrings needed a new invariant: strings must not be duplicated
across sets. For example, the same string cannot be OS-specific and
platform-specific. Bazel rejects build files where this is the case
for deps.

This change introduces platformStringsBuilder. Strings are added to
sets, and information is stored about which set each string belongs
to. The Map and Clean methods are removed; when PlatformStrings is
built, it is already clean.

This affects option strings somewhat. Unprocessed groups of options
are stored as strings with OptSeparator between each option. rules is
now responsible for splitting and quoting these.

This means order is no longer preserved for option strings (they will
be sorted when PlatformStrings is built). We couldn't actually
guarantee order before: for example, if a file contained an
arch-specific #cgo directive, followed by an OS-specific #cgo
directive, they would already appear in the wrong order.

Fixes #28