Vendor github.com/pelletier/go-toml at v1.0.1 (#71)
When rules_go builds Gazelle at master for use with go_repository, it fails because the dependency on github.com/pelletier/go-toml is missing. Adding go-toml to go_repository_tools (which downloads an extracts an additional tarball) seems way too gross, so I'm vendoring it instead.

This is temporary. As soon as we remove go_repository, we will be able to switch back to an external repository.