Switch to normal vendoring for go-toml
local_repository paths are relative to the directory containing the
main WORKSPACE file. Gazelle declared a local_repository in the
gazelle_dependencies macro for go-toml with a relative path pointing
to the vendor directory. This works inside the bazel_gazelle
workspace, but it breaks other repositories that import and use the
gazelle_dependencies macro.