update-repos: add or update repository rules by import path
The update-repos command can now add or update go_repository rules by
import path. To use:

    gazelle update-repos example.com/repo1 example.com/repo2

This will automatically discover the remote, VCS, and latest commit
for each repository containing the named packages. If a repository
already exists in WORKSPACE, it will be updated in place.

This functionality replaces wtool, which will be deleted soon after
this lands.

Related #56