internal/overlay_repository.bzl: add experimental overlay rules (#119)
These are drop-in replacements for http_archive and git_repository
with the added feature that you can copy in an "overlay" set of files
using custom names. The overlay is specified as a dictionary mapping
labels to relative paths in the destination repository.

In the future, these rules will be an alternative to
go_repository. Build files may be pre-generated using Gazelle, and
then copied into repositories using an overlay.

For now, these rules are not exposed in def.bzl and are not public yet.