Check that bazel_gazelle is loaded when fixing the WORKSPACE file
When `gazelle fix` or `gazelle update-repos` are run now, Gazelle
loads go_repository from @bazel_gazelle instead of @io_bazel_rules_go.
This is problematic for repositories that haven't updated to the new
Gazelle yet.

Gazelle now checks for @bazel_gazelle (or a gazelle:repo directive)
when modifying the WORKSPACE file. If it's needed but not present,
Gazelle will print an error.

Unfortunately, the repo can't be loaded automatically without a
hardcoding versions and sha256 sums, accessing the network, or Bazel
directory hackery.