New types for editing rules and load statements
This change introduces rule.File, Rule, and Load, which are
convenience types for editing build file syntax trees. These don't
provide a complete abstraction for build files: it's still
occasionally necessary to deal with expressions, and file parsing and
formatting are still required.

Rule and Load provide methods for adding, modifying, and removing
attributes and loaded symbols. Additionally, MergeRules and
SquashRules provide language agnostic merging and squashing
functionality.

The rest of Gazelle is not migrated to this new abstraction yet: that
will come in the next PR.