Set importmap in vendor to prefix concatenated with relative path (#207)
Previously, when entering a vendor directory, the importmap prefix was
set to the workspace name (read from WORKSPACE) concatenated with the
relative path in the repository. This was originally intended to be
just a unique value for disambiguating packages.

importmap is now used when formatting source file names. In
particular, it is used to place files in go_path. Gazelle should make
importmaps match what is expected in a GOPATH directory as much as
possible.

Unfortunately, this will result in some build file churn, but it seems
much better to fix this here than to add convoluted, error-prone logic
to go_path and anything else that uses importmap.

Fixes bazelbuild/rules_go#1509
Fixes bazelbuild/rules_go#1450
Fixes bazelbuild/rules_go#1406