Add code generation validator
This change adds a new "testing" directory that contains fake source
files that allow Gazelle authors to quickly iterate on designs. The test
library works by having real shaped directory structures then
automatically running Gazelle on it and comparing the output of that
against a checked in golden file.

Additionally there is a "update.sh" script that will run all the tests
in that directory and copy the generated BUILD.bazel file into the
appropriate source directory. This should keep the maintenance burden
for developers relatively low as well as provide a convenient way for
repo maintainers to see what would happen if they were to approve a PR
by inspecting what happens to the goldens.