internal/language/go: read prefix attribute from gazelle rule
Gazelle will now read the "prefix" attribute from a gazelle rule if
the prefix is not set on the command line or in a directive. This
removes an unintuitive sharp edge.

Also, update README.rst. Much earlier, the prefix attribute was
mandatory. This is no longer the case. The readme now advises people
to use the directive instead of the attribute.