fix-update: add -index, -r flags to control indexing and recursion
When the -r flag is true (it is by default), Gazelle will update
directories listed in positional arguments and subdirectories.

When the -index flag is true (it is by default), Gazelle will build an
index of libraries for dependency resolution.

When both -r and -index are false, Gazelle may visit only the
directories listed in positional arguments.