Create a gazelle_plugin rule
This rule allows one to define an additional language into the Gazelle
runner. By choosing this design, an author can write a generator that
plugs into gazelle but does not live in the Gazelle repo allowing for
independent iteration. Due to this independent iteration it is now also
possible to create a plugin that only works for a custom set of rules
that are, for example, developed in house. These in house and external
rules can then be interwoven with the rules_go provided plugins in
whatever order is necessary for the project.