repo: populate vcs field when importing Gopkg.lock with source (#410)
bazel-gazelle/cmd/fetch_repo command requires --vcs to be present when --remote is present. As a result, when there is source field in dep lock file, the vcs field should be populated for the resulting go_repository rules as well. Otherwise the subsequent bazel build will fail when it is calling gazelle to fetch external libraries.

Related #411