cmd/gazelle: save configurations for dependency resolution
runFixUpdate now saves the configuration for each visited subdirectory
in the corresponding visitRecord. The configuration is passed to
later language.Resolve calls for rules in that directory.

This allows the Go extension to use the prefix directive when deciding
how to resolve imports that couldn't be resolved with the index.

Fixes #412