Do not check if go_prefix is a prefix of import path
There are legitimate cases when an import path can have the current
repository's go_prefix as a prefix and yet can belong to a different
repository.

For example, v.io is github.com/vanadium/core, and v.io/x/lib is
github.com/vanadium/go.lib.

When doing external dependency resolution, we should rely on the result
from vcs.RepoRootForImportPath if the import path is not in the
resolver's rule index. For vendored dependency resolution, we can keep
the existing behavior.