language/go: resolve missing imports under current prefix as external (#409)
If -index=true, and we can't resolve an import using the index, we'll now resolve it as external or vendored, even if the import is under the current prefix.

Previously, we resolved all imports under the current prefix locally. This made it impossible to resolve to external repositories whose import paths started with the current prefix. For example, if the current prefix was v.io, Gazelle would resolve v.io/x/lib locally instead of externally. The new behavior only applies when -index=true in order to preserve autogazelle functionality.