Include hidden directories during walking (#415)
As discussed in #164, Go allows to import packages under hidden directories. If that happens, there is no way to generate build files for those packages. On the other hand, if there are unbuildable Go files in hidden directories, but Gazelle generates build files for them, it is easy to exclude them with # gazelle: exclude directives, the same strategy with "testdata"

Our scenario: in some of our projects, we check in the generated code under .gen directories. Those directories should be treated as regular Go package.

Fixed #164