Add ability to update proto.csv by results of bazel query.
This is a prerequisite to go_googleapis -> com_google_googleapis repository migration.

Since `com_google_googleapis` repository is Bazel-aware and declares its own `proto_library` and `go_proto_library` targets, we cannot just traverse the file tree of the repository and assume that there will be a corresponding `proto_library` and `go_proto_library` target for each found .proto file in a predefined format (the existing `BUILD.bazel` files in `com_google_googleapis` decide what the actual targets are and if they exist or no).

This PR provides an ability to construct protos.csv file based on the results of bazel query command, which outputs the existing `proto_library` and `go_proto_library` targets in `com_google_googleapis` repository.

To run the bazel query switch to `googleapis` repo directory and run:
    `bazel query 'kind("proto_library rule", //google/...)' --output xml --experimental_enable_repo_mapping > /path/to/googleapis/googleapis.xml`

To generate `protos.csv` from the generated `googleapis.xml`, switch to `language/proto/gen` directory (in gazelle workspace) and run:
    `go run update_proto_csv.go update_proto_csv_from_query.go -proto_csv=../proto.csv -googleapis_xml=/path/to/googleapis/googleapis.xml`

To generate `protos.csv` in the "old way", switch to `language/proto/gen` directory (in gazelle workspace) and run:
    `go run update_proto_csv.go update_proto_csv_from_query.go -proto_csv=../proto.csv -go_googleapis=/path/to/googleapis/repo`