map_kind directive: fix bug in handling of packages lacking a BUILD file
This previously assumed a *rule.File would be present for every kind mapping,
which is not the case for packages that lack a pre-existing BUILD file. This
change indexes the mapped kinds by the package relative path instead, with ""
denoting "ignore mapped kinds". This also adds a test to identify the problem
and verify the fix.

Fixes #469