Clarifies "ambiguous import" message
The old error message is not very useful: it is difficult to scan
visually, once you scan it it is not clear what it means, it offers no
suggestions at all what to do to fix the issue, and is very difficult to
google a solution for.

The new error is somewhat more verbose, but uses clearer language
(where the issue is comes first, details later) and offers clear ways
to resolve the issue.

Fixes #478