language/go: experimental module mode
This change adds a new flag to the Go extension,
-go_experimental_module_mode. It is off by default, but is turned on
when go_repository runs in module mode or when a go.mod file is
found. In the future, this flag will be removed, and we'll always run
in module mode.

When enabled, this flag changes how remote dependency resolution
works. Instead of using golang.org/x/tools/go/vcs (which does not
know about modules), we'll use "GO111MODULE=on go list <path>" in a
temporary directory. No special cases are enabled, but known
repositories from WORKSPACE and -known_repo arguments will be
observed.

Fixes #473