Add -to_macro option to update-repos (#480)
The to_macro option tells Gazelle to write repository rules into a .bzl macro function rather than the WORKSPACE file. The user can then load the macro into their WORKSPACE and call it from there. This helps with the maintainability of the WORKSPACE file, especially for projects with many go_repository rules.

The format of the option is -to_macro=macroFile%defName

Fixes #301