go_repository: replace attribute (#502)
go_repository now provides a "replace" attribute. This allows an
alternative module to be downloaded. The "version" and "sum"
attributes are used on the replacement, not the original
module. Gazelle is run as if it were on the original module though, so
import paths and labels will be the same.

Fixes #488