go_repository_tools: use generated file with list of srcs
Previously, changing a source file caused a slow re-evaluation of
go_repository_tools. bazel would start evaluating the rule, run the
script to list source files, resolve a source file that had changed,
then restart. Restarting causes the script to be rebuilt, which is
slow.

With this change, we run the script ahead of time to generate a .bzl
file with the list of sources. The list is an implicit dependency of
go_repository_tools. We still run the script once to verify the list
is up to date.