language/go, go_repository: submodule visibility
A Go package in an internal directory is visible to packages in
subdirectories of the parent of the internal directory. This
visibility is based on package path, which transcends module
boundaries. This means that a parent module may expose internal
packages to a submodule.

This doesn't map well to Bazel: each module is a separate workspace
(defined with a go_repository rule). Previously, we used labels like
"//:__subpackages__" for internal visibility, but now we also need
to add a label like "@com_google_cloud_go_logging//:__subpackages__"
for each submodule.

This PR makes several changes in order to do so:

* language/go now supports a directive,
  '# gazelle:go_submodule repo path', which maps a workspace name to a
  module path. The named module path should have the module path of the
  current workspace as a prefix.
* language/go now generates visibility attributes for internal
  packages that include references to submodule repositories when
  visibility should be allowed by Go internal packaging rules.
* go_repository now supports a build_submodules attribute, which
  accepts a string dictionary, mapping workspace names to module
  paths. This is translated into '# gazelle:go_submodule'
  directives. It's a separate attribute from build_directives so that
  it can be mergeable.
* When update-repos imports a go.mod file, it determines what modules
  are prefixes of other modules and adds build_submodule attributes to
  generated go_repository rules as needed.

Fixes #619