language/go: simpler submodule visibility (#648)
This change supercedes #637, which was reverted.

In #647, Gazelle now reads the repository configuration file (set with
-repo_config; WORKSPACE by default) and adds repository rules to the
Repos list, which all languages may access.

When language/go is configured, it reads Repos and makes a list of
submodules by looking for go_repository rules with importpaths that
have the current module path as a prefix. This is simpler than the
previous approach, which required explicitly adding build_submodule
attributes to go_repository.

When Go libraries and binaries are generated in internal directories,
the submodule list is used to add additional visibility entries.

Fixes #619.