repo: repository directive (#651)
A repository directive specifies a repository rule that Gazelle should know about. The directive can be repeated multiple times and can be declared from within a macro definition that Gazelle knows about. At the very least the directive must define a rule kind and a name 
attribute, but it can define extra attributes after that

Fixes #132