update-repos: add repository macro call and directive if missing (#695)
In update-repos, if the -to_macro flag is used on a macro that isn't
mentioned in a '# gazelle:repository_macro' directive in WORKSPACE,
add a directive.

Gazelle needs to be able to read repositories from WORKSPACE for
accurate and efficient dependency resolution. Macros must be declared
with '# gazelle:repository_macro'. It's a common mistake to use
'-to_macro', then call the macro without declaring it. This change
removes that sharp edge by adding a call with a directive
automatically. If a call is already present without a directive, a
directive will be added, but if a directive is present without a
call, it will be left alone.

Updates #689
Fixes #680
Updates #665