switch to using doublestar.Match
doublestar.Match is equivalent to path.Match whilst doublestar.PathMatch
is equivalent to filepath.Match. The main difference is that
doublestar.Match considers / as the file separator whilst PathMatch
expects path separator to be os-dependant. We want to remain OS agnostic
in here since the BUILD files are constant regardless of the operating
system.