language/go: don't emit select expressions for unsupported platforms
language/go now loads the io_bazel_rules_go version when Configure is
called for the root directory. Previously, we loaded the version in
cmd/gazelle/fix-update.go, which doesn't make sense when the Go
extension isn't being used. As before, we'll warn if the version can't
be loaded or is too old when the "fix" command is used.

When building select expressions, we'll ignore os and arch tags and
combinations that aren't supported by the current version of
rules_go. The oldest supported version of rules_go is 0.20.0, but
several new platforms were added in 0.23.0.

Verified manually. Tests can't easily change the version of rules_go.

Fixes #780