support dependencies on mapped kinds
Previously, rules with a mapped kind added to the RuleIndex would not
be importable.

The MetaResolver records mapped kinds generated by recorded associated
with the package. When looking up an import, we do not have the
associated package. This change addresses that by storing the language
name on the ruleRecord itself, obviating the need to look up the
Resolver.

The additional test for this scenario fails prior to this change and
succeeds as a result of it.

Fixes #943