Allow ListRepositories to load macros not in Workspace (#1053)
The PR allows a user to load other repository macros from within a top-level repository macro, so that it does not have to be added to the Workspace file. These macros with a second layer are identified by prepending a "+" to the beginning of the path in the repository_macro directive:

e.g. # gazelle:repository_macro +repos.bzl%go_repositories would load repos.bzl%go_repositories and any other files loaded from within the macro.

This is needed because the functionality already exists and is legal in bazel: A go_repository( ... ) rule can be added to a macro, loaded by another macro, and loaded by the workspace. However, gazelle will not resolve this repository, and create a duplicate rule, thereby creating incorrect/duplicate Starlark content.

In very large repositories, this is important when dependency management needs to happen in many different file spaces.
This change uses a highly conservative opt-in approach, although since it is legal in Bazel, perhaps this "opt-in" requirement can be removed in the future. With this implementation, ListRepositories will correctly list all of the go_repository( imports in that file, whether directly or loaded from another macro, as Bazel would similarly load.

For all existing uses (macro files existing as references only from Workspace file), no changes would be made. Only in cases where a user specifically decides to load another macro, and identify this desire in the gazelle directive.