Stardoc generates repository.md
This is a mirror of the content in repository.rst, allowing us to manually inspect the delta between them.
As soon as we are satisfied that the generated content is correct, we'll delete repository.rst and update any links to it.

Technique: copied doc strings out of existing rst file into the starlark files. Transcribed rst syntax to markdown.