cmd/autogazelle: improve walk behavior to only watch configured directories
Fixes #759.

The `autogazelle` command would previously watch all directories in the entire
workspace. This can be extremely bad. In the `cockroachdb` repo, this meant
that even with an extremely 100k file-descriptor limit, the tool could not
watch everything and at 1M, the terminal crashed. Instead, we have the power
to listen to gazelle directives and use the `walk` functionality.