Report error details from `go mod download`
Currently if there are errors from `go mod download`, e.g. if a URL
couldn't be resolved or connected to, the output is swallowed because we
ignore the JSON which describes the error.

This change attempts to best-effort parse the output JSON to find error
description strings, and attach them to the error we report to users.