fix(lang/proto): include imports from different targets
Fixes an issue where dependencies would be missed in the case where an
import was of a proto file in the same directory but from a different
target, such an when using `proto_group`. This would cause Gazelle to
generate `proto_library` and `go_proto_library` targets that could not
be built.

Fixes #1219