Respect .bazelignore
Fixes #318

The code for this re-purposes the original PR #1022 but makes a few
modifications for performance and correctness.

First we only load the .bazelignore once when `rel == ""` since that means
we're in the root; the previous implementation seemed to do the same load
for every directory visited. Next we only match files relative to repo
root to have parity with how Bazel itself parses the ignore file.