Don't exclude spaces from the label name regex
In parsing the output of `bazel query`, we found that some files with
spaces in their names are included in the output.

It seems reasonable that all labels that come out of a `bazel query`
should be parsable by this `Label` type; let's just use the raw regex
from the Bazel documentation.