Expose rule.Rule as Expr
This allows for richer information gathering about existing rules in
BUILD files. In particular, it allows for calling `bzl.Walk` on an
entire Rule, whereas currently only individual args and attrs values
are exposed.

An example of where this is insufficient is that right now there's no
way to read the comment in this rule:

```starlark
filegroup(
    name = "fg",
    # This comment is important
    srcs = ["foo"],
)
```

because the existing API only surfaces the attr value as an Expr
(`["foo"]`), and not the `AssignExpr` (`srcs = ["foo"]`), to which this
comment  is actually attached.