Record where dependencies come from
This allows rules and aspects to determine where a particular
Go dependency has come from. This can be useful when (eg)
generating SBOMs directly from Bazel projects.

The approach of storing the information in `tags` is copied
from `rules_jvm_external`, which uses a similar technique to
store maven coordinates.

Build files are amended using Buildozer, since there is no
way to be sure whether or not a specific `go_repository` had
build files before being created. Where Buildozer is not
available, a warning is printed and no version information is
added to the tags of the generated libraries.