Modify gen_known_imports to directly create structs
The previous system, which generated calls to label.New(), was creating Go binaries that were over 80MB in size due to all the garbage collector metadata needed.

This PR instead generates struct literals, which reduces that to ~10MB total (in our configuration). There are probably still better ways to include the known imports maps, but this resolves the immediate issue for us.