gazelle: add CPU profiling to aid troubleshooting
As Gazelle becomes the defacto tool to generate BUILD files for several
languages in Bazel ecosystem, we gain more contributors and extension
authors from a diverse set of background and experience.

However, Gazelle itself is complicated and not all the contributors are
familiar with the Go ecosystem. This led to difficulty in developing
language extensions and troubleshooting edge cases.

Provides a way to turn on Go profiling(1) is a great way to help with
that. Go's pprof tool is well-documented and CLI friendly. Rules author
could simply turn on the CPU profile capture to analyze it and figure
out where is the bottleneck, how does the call stack look like.

Add environment variables `GAZELLE_CPU_PROFILE` to collect Go CPU
profile for later analysis.

(1): https://go.dev/blog/pprof