bzlmod: Allow overriding default Go module configuration (#1456)
Introduce a new "gazelle_override" tag for the "go_deps" module
extension, allowing callers to specify directives to Gazelle
overriding its normal behavior when preparing an external Bazel
repository for a given Go module. Rather than allowing tuning the
generated "go_repository" targets with just its two
"build_file_proto_mode" and "build_naming_convention" attributes,
instead generalize the overriding facility to work with Gazelle
directives, which are strings with a keyword and an accompanying
argument for those keywords that accept a parameter.

As before with the "go_repository" rule, one can override both the
BUILD file naming convention and how Gazelle generates rules for
Protocol Buffers files for each per-Go module repository, but those
are just two of the larger set of directives honored by Gazelle.

Each "gazelle_override" tag occurrence must refer to a Go module
defined by the containing Bazel module; it is not possible to override
characteristics of a Go module defined by another Bazel
module. Enforcing this correspondence ensures that the overridden
characteristics are appropriate for the given version of the Go module
defined by the containing Bazel module, as different versions of the
Go module may warrant different characteristics. Furthermore, one may
only use the "gazelle_override" tag in a root Bazel module.

In order to reduce the burden on consuming modules, honor a default
set of Gazelle directives for Go modules known to require them.

Here we change the interface to two existing facilities:
- Permit use of at most one "go_deps.from_file" tag
- Deprecate two "go_module" tag class attributes

  Push callers to favor use of the "gazelle_overrides" tag class's
  "build_naming_convention" and "build_file_proto_mode" attributes
  instead of the same-named attributes on the "go_module" tag class,
  but continue to honor the latter for now.

* Let Gazelle find the repo config with Bzlmod

With Bzlmod, there is no WORKSPACE file for Gazelle to learn the naming
conventions of external repos from, so we point it to the synthetic repo
config generated by the go_deps extension.

Co-authored-by: Steven E. Harris <seh@panix.com>

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>