Option to include diff header line
`git apply` works for existing files only when there is a "git diff"
header line for each file that begins with `git --diff`. Without this
header line, we get `error: new file ... depends on old contents`

It may be reasonable to give Gazelle an option to output such lines to
play nice with git.

In general, `diff -u` also outputs a similar header line, although
`patch` does not seem to care as long as the patch conforms to the
unified diff format.