Make known imports package vars public
Gazelle has a "known imports" feature that operates during the resolve phase.
First, the overrides are checked.  If an override is not found, known imports are
checked.  If a known import is not found, the index is checked.

There is no way to opt-out of the known imports for "well known" types
from @com_google_protobuf and @go_googleapis in most indexing modes.

This makes it not possible to use gazelle to index the protos in those repositories
(in file mode for example) because the resolve data should come from the index.
In that case, override rules don't make sense (one would need to pre-index the repo
being indexed!); the hardcoded ones correspond to -mode=package so they are not
appropriate for file mode (or other modes).

This PR makes the known imports related variable public, so gazelle extension
authors have full control over the known imports and can set it at initialization time.
In the case desribed above, an empty known import map is the correct one to resolve
from the index.