Fix partial execution with mapped kinds (#1680)
Previously, if mapped kinds were in use (i.e. `go_library` -> `go_custom_library`),
and `gazelle update` was being run on just one directory, Gazelle's
indexing system would fail to correctly index build rules of the mapped-to
kind in directories that were not being re-built.

This is because the logic to handle mapped kinds was triggered only when
Generate() was run, not for Imports().

This PR solves the issue with the smallest fix. The whole kind-mapping
system is honestly kinda messy, and I didn't want to propose a major
refactor.

Co-authored-by: Eric Miller <eric@skydio.com>