rule: add accessor for attribute comments (#1700)
This allows user to attach comments like "# do not sort" to the
assignment expressions maintained by rule.go.

rule.Format() calls into buildtools/build.Format() which in turn calls
Rewrite(). Rewrite() by default sorts certain label arguments like
"deps" and "hdrs". To preventing this sorting it is possible to provide
the comment "# do not sort". However this comment needs to be attached
to the assignment expression.
In rule.go the assignment expressions are not exposed to the user,
making it impossible to attach the comment to it.

See https://github.com/bazelbuild/bazel-gazelle/issues/237#issuecomment-399476945