[Gazelle] Fix Duplicate Load Bug (#1841)
If two extensions both import Kinds from the same file, and no load
statement for those Kinds is already in the BUILD.bazel file, then
Gazelle will add the load statement twice.

In this commit, we fix the issue by updating the tracked loads when a
new load is inserted into the index. This will cause future iterations
to update the existing load instead of creating a duplicate.