[Proto] Keep track of the names of Services, Messages, and Enums
In order to construct a fully qualified Java import from the proto file,
we need to know the values of the "java_package" option as well as the
services, messages, and enums in the proto file. The "java_pacakge"
option is already recorded. In this commit, we add support for parsing
the names of the services, message, and enums of a proto file and
tracking them in the FileInfo struct.