[Proto] Require space between 'service' and service name in regex matching
Prior to this commit, the regex for determining whether a file had
services did not require a space between the word "service" and the
service name. This could lead to false alarms, like "message
serviceABC". In this commit, we update the regex to only match if there
is at least one space between service and the service name.