refactor: compute transitive embeds during indexing (#1868)
**What type of PR is this?**
refactor

**What package or component does this PR mostly affect?**
all

**What does this PR do? Why is it needed?**
Refactor working toward serializing+caching of the rule index.

The `ruleRecord` should be as close as possible to what a language
returns. All computed data such as that for indexing should be computed
from the `ruleRecord` and persisted elsewhere.

**Which issues(s) does this PR fix?**

Ref: #1181
Some ideas are similar to #1182 while trying to reduce the scope of the
change and diff.

**Other notes for review**

IMO this is a reasonable cleanup on its own, without even taking
serializing+caching into account 🤷